handlers_refactor split input_handlers.py into event_handlers directory, one handler per file documented and annotated all handlers moved keys to a constants file... constants refactor is next
