+++ src/components/fighter.py
@@ -2,7 +2,7 @@ from components.base import BaseComponent
 from constants import colors, move_elevations
 from entity import Actor
 from enums import GameStates, RenderOrder
-from input_handlers import GameOverEventHandler
+from event_handlers.player_dead import GameOverEventHandler
 
 
 class Fighter(BaseComponent):
